requirements.txt
clickhouse-connect==0.4.8
fastapi==0.89.0
faust-streaming==0.10.1
kafka-python==2.0.2
passlib[bcrypt]==1.7.4
tortoise-orm[asyncpg]==0.19.2
uvicorn[standard]==0.17.6
